+++ packages/shaam6111-generator/src/generators/index.ts
@@ -1,3 +1,3 @@
 export * from './generate-report.js';
-export * from './header.js';
-export * from './report-entries.js';
+export * from './generate-header.js';
+export * from './generate-report-entries.js';
